refactor: generalise plotting argument checks
Commands no more call plotting code directly. They only craft plotting
code and return it to the main function. Main function checks if plot
was requested and calls it. Unfortunately plotting code is highly
customised so it is not enough just to return data from commands and
plot it with generic code.